+++ docs/CHANGELOG.md
@@ -1,3 +1,7 @@
+---
+robots: noindex, follow
+---
+
 # Changelog
 
 All notable changes to this project will be documented in this file.